Add a table-based view for the performance data (#6849)
* Fix issue where the bottom of the panel was cut off from view

This patch changes the details panel's container to use padding
rather than margin. With margin, the size of the container was using the
outer size, and cutting off the bottom of the contents. With padding,
the contents are correctly sized for the space and nothing is cut off.

A more correct fix would be to remove the margin and padding on the
outer container, and then ensure that the contents of each details panel
could determine its own padding. This would ensure the scroll bar is all
the way to the right and didn't cover any content. However, this would
be a riskier change, and would need each panel tested.

* Add a table-based view for the performance data

This includes the unit and whether lower or higher is better. I felt a
bit confused looking at the results, and not knowing what the individual
values meant. This change also presents everything in a table format
which should make it easier to scan and understand.

I ran these changes by some users of this view, and handled some of
their early feedback. This included removing a graph that wasn't very
helpful. There might be some opportunities in the future to add more
details to the columns to help in making navigating the data easier.
However, this is not a replacement for the comparison view, but instead
the motivation is for making the individual test runs easier to
understand.

* Address the review comments on preferring bootstrap classes over custom styles

I also added a font size 16 class, which is equivalent to an <h6> font
size in the Bootstrap system.