Bug 1679163 - Adjust tooltip placement to avoid it obscuring the target text
If the tooltip is long, it can obscure the target "view details" text,
which then triggers the mouseout event that hides the tooltip. This can
set up an oscilation between the tooltip showing and hidding in quick
succession, which manifests as the user's cursor rapidly flickering.

Moving the tooltip to be above the target text seems to solve this
problem as it reduces the chance that the target text will be obscured.