Bug 1617835 - optimize time to switch to view showing single push (#7222)
The object holding the mapping of job IDs to the jobs themselves got iteratively
regenerated - the script walked over all jobs currently shown (with
Array.prototype.reduce) and for each job of the push which shall be shown, a new
object got generated based on the old one + the new job id and data. Based on
testing with mozilla-central and the default 10 pushes shown, this translates
into ~90k jobs and slow page warnings (execution up to 46s measured). The new
code takes one object and sets the id and data pairs iteratively in 100-200ms.