Bug 1722731 - remove filtering by 'machine name' (#7226)
Bug 1450045 removed the machine_name data from the frontend cache and the
ability to filter by 'machine name' has been broken since.

Checking the results of tasks which have been run a specific machine should be
done with Taskcluster. Treeherder would also only show the runs for the
currently shown tree, to which machines are not bound.