Bug 1842924 - Update deletion constraints for performancedatumreplicate.
This patch adds a migration to update the foreign key `on_delete` constraint for PerformanceDatumReplicate. This happens because Django doesn't apply the cascade operations at the database-level (only the ORM-level). To resolve this, we have to find the existing foreign key constraint label using a database viewer, and update it to add `ON DELETE CASCADE`. Follow this Django bug for progress on this issue: https://code.djangoproject.com/ticket/21961