Bug 1867010 - Integrate replicates into the Graphs View. (#7887)
This patch adds a button to enable using replicates in the graphs view. The button is made available all the time as the cases for where replicates are available are not limited to a single branch anymore after PR #7886 lands. It also changes the API to return replicates for the specific case that the graphs view querys for (all_data).