Bug 1879316 - Use the absolute value of the delta in the alert threshold check. (#7909)
This patch fixes an issue where the absolute alert threshold was negative on an improvement, and this was checked against the alert threshold. This meant that improvements would never have an alert generated for them since they were always negative. With this change, improvement alerts will now be produced for tests using the absolute threshold.