Bug 1910864 - Return None for profiler urls instead of "N/A"
Perfherder frontend has a check if the profiler urls are null, but it
doesn't check if it's "N/A" or not. That's why this was causing us to
see all the invalid links in the perfherder when user clicks on "Copy
Summary" or file a bug directly through perfherder. By changing this to
`None`, this API endpoint will return `null` to the frontend, which can
successfully understand and omit.