Bug 1910864/1887138 - Fix how we handle profiler urls on the perfherder backend (#8156)
* Bug 1910864 - Return None for profiler urls instead of "N/A"

Perfherder frontend has a check if the profiler urls are null, but it
doesn't check if it's "N/A" or not. That's why this was causing us to
see all the invalid links in the perfherder when user clicks on "Copy
Summary" or file a bug directly through perfherder. By changing this to
`None`, this API endpoint will return `null` to the frontend, which can
successfully understand and omit.

* Bug 1887138 - Refactor how we compute the profile urls on the backend

There were two issues with how we compute the profile urls in the
backend:

1. When we encounter an error, we were returning an error string. This
   is not good because frontend doesn't know how to handle the error
   strigs and it thinks that they are simply urls. Then it was trying to
   append the error strings on the back of profiler urls, which was
   causing invalid urls in the perf alerts.
2. We had a caching mechanism for the taskcluster metadata. But this
   caching mechanism wasn't really caching things by adding task id to
   the cache key. This was resulting a single cache for the whole "push"
   or multiple pushes. We were caching the latest metadata, and use that
   last metadata for all the alerts, which is wrong because all
   individual alerts might have differnet metadata. The good thing is,
   the taskcluseter metadata was already available to us inside the
   `alert` variable since we computed it prior this function. So we
   didn't have to do anything special, just read that from the
   dictionary.

There is one remaining piece I would like to fix still, which is how we
cache `tc_root_url`. But that's actually okay for now because a single
push should only have one `tc_root_url`. And I'm not so sure how to
pass that information to the other function where we genereate the
profile urls.