Bug 1919588 - Add new alert tables, along with new fields to existing table. (#8470)
This patch adds a new alert table for testing new detection techniques. Some things are changed with the alert table definitions to be able to reuse a lot of the code in the new tables. At the same time, one of the methods (autodetermine_status) is modified to be more generic and usable by subclasses.

Two new fields are added called confidence, and detection_method. The confidence field is a more generically termed field similar to the t_value field but isn't specific to T-test. The detection_method field will be used to store the name of the detection method used to produce the alert.

Additionally, a new table for telemetry alerts is added here, and makes use of the *Base classes that were added.