Bug 1942860 - Add prototype for telemetry alerting to perf_sheriff task.
This patch adds the ability to produce telemetry alerts. It runs through the perf_sheriff task in newrelic. This is a prototype so all of it is hidden behind a catch-all try/except so it doesn't impact any existing capabilities (this include the imports).

The alerts are detected through mozdetect. They are then added into the PerformanceTelemetryAlert, and PerformanceTelemetryAlertSummary models, and associated to specific mozilla-central pushes.