Bug 1947517 - Rework the alert comment to make information more accessible. (#8583)
This patch changes the alert comment #0 to make some information more accessible for developers. It moves the backout message to the top, and follows it with a link to some documentation. At the same time, some additional info is provided for where to reach out to people for help.

The wiki link to the sheriffing docs is also removed since it hasn't been useful for developers. The backout mention is also changed to mention that we're only looking for an acknowledgment, and an investigation to begin within 3 business days (instead of a plan).