Bug 1964128 - Enable monitor-only alerting for performance tests with email notifications. (#8670)
This patch adds the ability to produce performance alerts on tests that are only for monitoring purposes, and will not be sheriffed. This includes a new interface to view them on the Perfherder page (Monitored Alerts). At the same time, this patch adds the ability to send emails for all alerts (including sheriffed ones) to anyone that has an email specified in the new alertNotifyEmails field.