Bug 1999984 - Ignore telemetry alerts with no bug numbers during housekeeping. (#9069)
This patch fixes an issue in telemetry alerting that occurs when an alert with a bug number is associated with an alert summary that contains alerts without a bug number. It adds some checks to prevent a None bug from being modified, and prevents the alert without a bug from being picked up by housekeeping.